adding auth to admin & seller and fixing filter by name

+++ controllers/department.js
@@ -10,10 +10,16 @@ const getAllDepartments = async (req, res, next) => {
 };
 
 const AddnewDep = async (req, res, next) => {
-  const departmentToBeSaved = req.body;
   try {
-    const addededDepartment = await departmentModel.create(departmentToBeSaved);
-    res.status(201).json(addededDepartment);
+    if (req.role === "admin") {
+      const departmentToBeSaved = req.body;
+      const addededDepartment = await departmentModel.create(
+        departmentToBeSaved
+      );
+      res.status(201).json(addededDepartment);
+    } else {
+      res.status(500).json("you not admin");
+    }
   } catch (err) {
     res.status(422).json({ message: err.message });
   }
@@ -30,23 +36,31 @@ const getDepById = async (req, res, next) => {
 };
 
 const updateDepById = async (req, res) => {
-  const id = req.params.id;
-  const obj = req.body;
   try {
-    let updatedDep = await departmentModel.findByIdAndUpdate(id, obj, {
-      new: true,
-    });
-    res.json(updatedDep);
+    if (req.role === "admin") {
+      const id = req.params.id;
+      const obj = req.body;
+      let updatedDep = await departmentModel.findByIdAndUpdate(id, obj, {
+        new: true,
+      });
+      res.json(updatedDep);
+    } else {
+      res.status(500).json("you not admin");
+    }
   } catch (err) {
     res.status(422).json({ message: err.message });
   }
 };
 
 const deleteDepartment = async (req, res) => {
-  const id = req.params.id;
   try {
-    await departmentModel.findByIdAndDelete(id);
-    res.json("Department Deleted Successfully");
+    if (req.role === "admin") {
+      const id = req.params.id;
+      await departmentModel.findByIdAndDelete(id);
+      res.json("Department Deleted Successfully");
+    } else {
+      res.status(500).json("you not admin");
+    }
   } catch (err) {
     res.status(422).json({ message: err.message });
   }
